`struct WithOffset<T>`: Move `impl {Add,Sub}{,Assign}`s of `Rav1dPictureDataComponentOffset` to a generic type.
