+++ include/dav1d/picture.rs
@@ -26,16 +26,13 @@ use crate::src::error::Dav1dResult;
 use crate::src::error::Rav1dError;
 use crate::src::error::Rav1dError::EINVAL;
 use crate::src::error::Rav1dResult;
+use crate::src::with_offset::WithOffset;
 use libc::ptrdiff_t;
 use libc::uintptr_t;
 use std::array;
 use std::ffi::c_int;
 use std::ffi::c_void;
 use std::mem;
-use std::ops::Add;
-use std::ops::AddAssign;
-use std::ops::Sub;
-use std::ops::SubAssign;
 use std::ptr::NonNull;
 use std::sync::Arc;
 use to_method::To as _;
@@ -400,79 +397,7 @@ impl Rav1dPictureDataComponent {
     }
 }
 
-#[derive(Clone, Copy)]
-pub struct Rav1dPictureDataComponentOffset<'a> {
-    pub data: &'a Rav1dPictureDataComponent,
-    pub offset: usize,
-}
-
-impl<'a> AddAssign<usize> for Rav1dPictureDataComponentOffset<'a> {
-    #[cfg_attr(debug_assertions, track_caller)]
-    fn add_assign(&mut self, rhs: usize) {
-        self.offset += rhs;
-    }
-}
-
-impl<'a> SubAssign<usize> for Rav1dPictureDataComponentOffset<'a> {
-    #[cfg_attr(debug_assertions, track_caller)]
-    fn sub_assign(&mut self, rhs: usize) {
-        self.offset -= rhs;
-    }
-}
-
-impl<'a> AddAssign<isize> for Rav1dPictureDataComponentOffset<'a> {
-    #[cfg_attr(debug_assertions, track_caller)]
-    fn add_assign(&mut self, rhs: isize) {
-        self.offset = self.offset.wrapping_add_signed(rhs);
-    }
-}
-
-impl<'a> SubAssign<isize> for Rav1dPictureDataComponentOffset<'a> {
-    #[cfg_attr(debug_assertions, track_caller)]
-    fn sub_assign(&mut self, rhs: isize) {
-        self.offset = self.offset.wrapping_add_signed(-rhs);
-    }
-}
-
-impl<'a> Add<usize> for Rav1dPictureDataComponentOffset<'a> {
-    type Output = Self;
-
-    #[cfg_attr(debug_assertions, track_caller)]
-    fn add(mut self, rhs: usize) -> Self::Output {
-        self += rhs;
-        self
-    }
-}
-
-impl<'a> Sub<usize> for Rav1dPictureDataComponentOffset<'a> {
-    type Output = Self;
-
-    #[cfg_attr(debug_assertions, track_caller)]
-    fn sub(mut self, rhs: usize) -> Self::Output {
-        self -= rhs;
-        self
-    }
-}
-
-impl<'a> Add<isize> for Rav1dPictureDataComponentOffset<'a> {
-    type Output = Self;
-
-    #[cfg_attr(debug_assertions, track_caller)]
-    fn add(mut self, rhs: isize) -> Self::Output {
-        self += rhs;
-        self
-    }
-}
-
-impl<'a> Sub<isize> for Rav1dPictureDataComponentOffset<'a> {
-    type Output = Self;
-
-    #[cfg_attr(debug_assertions, track_caller)]
-    fn sub(mut self, rhs: isize) -> Self::Output {
-        self -= rhs;
-        self
-    }
-}
+pub type Rav1dPictureDataComponentOffset<'a> = WithOffset<&'a Rav1dPictureDataComponent>;
 
 impl<'a> Rav1dPictureDataComponentOffset<'a> {
     pub fn stride(&self) -> isize {
